fix(eng-analytics): rename test_health serializers module to suite_health
